finish graph convolutional network (gcn), graph attention network (gat), update lstm & n_gram & create graph_utils.py used by both gcn & gat network
